@@ -2,6 +2,7 @@
   "name": "rusty-butter-monorepo",
   "version": "1.0.0",
   "private": true,
+  "type": "module",
   "description": "Rusty Butter monorepo containing all packages",
   "scripts": {
     "build": "pnpm -r build",
@@ -12,7 +13,9 @@
     "lint:fix": "eslint . --ext .ts,.tsx,.js,.jsx --fix",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
-    "typecheck": "pnpm -r exec tsc --noEmit"
+    "typecheck": "pnpm -r exec tsc --noEmit",
+    "claude": "./scripts/claude-with-env.sh",
+    "claude:show-env": "./scripts/show-env.sh"
   },
   "engines": {
     "node": ">=16.0.0",
@@ -22,7 +25,9 @@
   "devDependencies": {
     "@rusty-butter/eslint-config": "workspace:*",
     "@rusty-butter/prettier-config": "workspace:*",
-    "eslint": "^8.56.0",
+    "@eslint/js": "^9.30.1",
+    "eslint": "^9.30.1",
+    "typescript-eslint": "^8.38.0",
     "prettier": "^3.2.5"
   }
 }
