@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "Shared ESLint configuration for Rusty Butter monorepo",
   "main": "index.js",
+  "type": "module",
   "files": [
     "index.js"
   ],
@@ -14,14 +15,14 @@
   "author": "Rusty Butter",
   "license": "MIT",
   "dependencies": {
-    "@typescript-eslint/eslint-plugin": "^7.0.0",
-    "@typescript-eslint/parser": "^7.0.0",
-    "eslint": "^8.56.0",
+    "@eslint/js": "^9.30.1",
+    "eslint": "^9.30.1",
+    "typescript-eslint": "^8.38.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3"
   },
   "peerDependencies": {
-    "eslint": ">=8.0.0",
+    "eslint": ">=9.0.0",
     "typescript": ">=5.0.0"
   }
 }
